@@ -1,7 +1,7 @@
 package org.bukkit.craftbukkit.block;
 
-import java.util.Random;
 import net.minecraft.core.BlockPosition;
+import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.EntityTypes;
 import net.minecraft.world.entity.animal.EntityBee;
 import net.minecraft.world.level.GeneratorAccessSeed;
@@ -29,7 +29,7 @@ public final class CapturedBlockState extends CraftBlockState {
         if (this.treeBlock && getType() == Material.BEE_NEST) {
             GeneratorAccessSeed generatoraccessseed = this.world.getHandle();
             BlockPosition blockposition1 = this.getPosition();
-            Random random = generatoraccessseed.getRandom();
+            RandomSource random = generatoraccessseed.getRandom();
 
             // Begin copied block from WorldGenFeatureTreeBeehive
             TileEntity tileentity = generatoraccessseed.getBlockEntity(blockposition1);
