@@ -11,7 +11,9 @@ import javax.annotation.Nullable;
 import net.minecraft.core.BlockPosition;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.world.level.GeneratorAccess;
+import net.minecraft.world.level.block.entity.SculkCatalystBlockEntity;
 import net.minecraft.world.level.block.entity.SculkSensorBlockEntity;
+import net.minecraft.world.level.block.entity.SculkShriekerBlockEntity;
 import net.minecraft.world.level.block.entity.TileEntity;
 import net.minecraft.world.level.block.entity.TileEntityBanner;
 import net.minecraft.world.level.block.entity.TileEntityBarrel;
@@ -23,6 +25,7 @@ import net.minecraft.world.level.block.entity.TileEntityBlastFurnace;
 import net.minecraft.world.level.block.entity.TileEntityBrewingStand;
 import net.minecraft.world.level.block.entity.TileEntityCampfire;
 import net.minecraft.world.level.block.entity.TileEntityChest;
+import net.minecraft.world.level.block.entity.TileEntityChestTrapped;
 import net.minecraft.world.level.block.entity.TileEntityCommand;
 import net.minecraft.world.level.block.entity.TileEntityComparator;
 import net.minecraft.world.level.block.entity.TileEntityConduit;
@@ -129,6 +132,8 @@ public final class CraftBlockStates {
                         Material.DARK_OAK_WALL_SIGN,
                         Material.JUNGLE_SIGN,
                         Material.JUNGLE_WALL_SIGN,
+                        Material.MANGROVE_SIGN,
+                        Material.MANGROVE_WALL_SIGN,
                         Material.OAK_SIGN,
                         Material.OAK_WALL_SIGN,
                         Material.SPRUCE_SIGN,
@@ -257,18 +262,12 @@ public final class CraftBlockStates {
                 ), CraftCampfire.class, CraftCampfire::new, TileEntityCampfire::new
         );
 
-        register(
-                Arrays.asList(
-                        Material.CHEST,
-                        Material.TRAPPED_CHEST
-                ), CraftChest.class, CraftChest::new, TileEntityChest::new
-        );
-
         register(Material.BARREL, CraftBarrel.class, CraftBarrel::new, TileEntityBarrel::new);
         register(Material.BEACON, CraftBeacon.class, CraftBeacon::new, TileEntityBeacon::new);
         register(Material.BELL, CraftBell.class, CraftBell::new, TileEntityBell::new);
         register(Material.BLAST_FURNACE, CraftBlastFurnace.class, CraftBlastFurnace::new, TileEntityBlastFurnace::new);
         register(Material.BREWING_STAND, CraftBrewingStand.class, CraftBrewingStand::new, TileEntityBrewingStand::new);
+        register(Material.CHEST, CraftChest.class, CraftChest::new, TileEntityChest::new);
         register(Material.COMPARATOR, CraftComparator.class, CraftComparator::new, TileEntityComparator::new);
         register(Material.CONDUIT, CraftConduit.class, CraftConduit::new, TileEntityConduit::new);
         register(Material.DAYLIGHT_DETECTOR, CraftDaylightDetector.class, CraftDaylightDetector::new, TileEntityLightDetector::new);
@@ -284,10 +283,13 @@ public final class CraftBlockStates {
         register(Material.JUKEBOX, CraftJukebox.class, CraftJukebox::new, TileEntityJukeBox::new);
         register(Material.LECTERN, CraftLectern.class, CraftLectern::new, TileEntityLectern::new);
         register(Material.MOVING_PISTON, CraftMovingPiston.class, CraftMovingPiston::new, TileEntityPiston::new);
+        register(Material.SCULK_CATALYST, CraftSculkCatalyst.class, CraftSculkCatalyst::new, SculkCatalystBlockEntity::new);
         register(Material.SCULK_SENSOR, CraftSculkSensor.class, CraftSculkSensor::new, SculkSensorBlockEntity::new);
+        register(Material.SCULK_SHRIEKER, CraftSculkShrieker.class, CraftSculkShrieker::new, SculkShriekerBlockEntity::new);
         register(Material.SMOKER, CraftSmoker.class, CraftSmoker::new, TileEntitySmoker::new);
         register(Material.SPAWNER, CraftCreatureSpawner.class, CraftCreatureSpawner::new, TileEntityMobSpawner::new);
         register(Material.STRUCTURE_BLOCK, CraftStructureBlock.class, CraftStructureBlock::new, TileEntityStructure::new);
+        register(Material.TRAPPED_CHEST, CraftChest.class, CraftChest::new, TileEntityChestTrapped::new);
     }
 
     private static void register(Material blockType, BlockStateFactory<?> factory) {
@@ -322,6 +324,16 @@ public final class CraftBlockStates {
     public static Class<? extends CraftBlockState> getBlockStateType(Material material) {
         Preconditions.checkNotNull(material, "material is null");
         return getFactory(material).blockStateType;
+    }
+
+    public static TileEntity createNewTileEntity(Material material) {
+        BlockStateFactory<?> factory = getFactory(material);
+
+        if (factory instanceof BlockEntityStateFactory) {
+            return ((BlockEntityStateFactory<?, ?>) factory).createTileEntity(BlockPosition.ZERO, CraftMagicNumbers.getBlock(material).defaultBlockState());
+        }
+
+        return null;
     }
 
     public static BlockState getBlockState(Block block) {
@@ -371,7 +383,7 @@ public final class CraftBlockStates {
         return factory.createBlockState(world, blockPosition, blockData, tileEntity);
     }
 
-    private static boolean isTileEntityOptional(Material material) {
+    public static boolean isTileEntityOptional(Material material) {
         return material == Material.MOVING_PISTON;
     }
 
