@@ -1,6 +1,7 @@
 package org.bukkit.craftbukkit.entity;
 
 import com.google.common.base.Preconditions;
+import net.minecraft.resources.MinecraftKey;
 import net.minecraft.world.entity.EntityInsentient;
 import org.bukkit.Bukkit;
 import org.bukkit.NamespacedKey;
@@ -61,11 +62,7 @@ public abstract class CraftMob extends CraftLivingEntity implements Mob {
 
     @Override
     public LootTable getLootTable() {
-        if (getHandle().lootTable == null) {
-            getHandle().lootTable = getHandle().getDefaultLootTable();
-        }
-
-        NamespacedKey key = CraftNamespacedKey.fromMinecraft(getHandle().lootTable);
+        NamespacedKey key = CraftNamespacedKey.fromMinecraft(getHandle().getLootTable());
         return Bukkit.getLootTable(key);
     }
 
