@@ -1,7 +1,8 @@
 package org.bukkit.craftbukkit.entity;
 
+import net.minecraft.core.Holder;
 import net.minecraft.world.entity.decoration.EntityPainting;
-import net.minecraft.world.entity.decoration.Paintings;
+import net.minecraft.world.entity.decoration.PaintingVariant;
 import org.bukkit.Art;
 import org.bukkit.block.BlockFace;
 import org.bukkit.craftbukkit.CraftArt;
@@ -16,7 +17,7 @@ public class CraftPainting extends CraftHanging implements Painting {
 
     @Override
     public Art getArt() {
-        Paintings art = getHandle().motive;
+        Holder<PaintingVariant> art = getHandle().getVariant();
         return CraftArt.minecraftToBukkit(art);
     }
 
@@ -28,12 +29,12 @@ public class CraftPainting extends CraftHanging implements Painting {
     @Override
     public boolean setArt(Art art, boolean force) {
         EntityPainting painting = this.getHandle();
-        Paintings oldArt = painting.motive;
-        painting.motive = CraftArt.bukkitToMinecraft(art);
+        Holder<PaintingVariant> oldArt = painting.getVariant();
+        painting.setVariant(CraftArt.minecraftToBukkit(art));
         painting.setDirection(painting.getDirection());
         if (!force && !getHandle().generation && !painting.survives()) {
             // Revert painting since it doesn't fit
-            painting.motive = oldArt;
+            painting.setVariant(oldArt);
             painting.setDirection(painting.getDirection());
             return false;
         }
