@@ -39,7 +39,7 @@ public class CraftMetaAxolotlBucket extends CraftMetaItem implements AxolotlBuck
         }
 
         if (tag.contains(ENTITY_TAG.NBT)) {
-            entityTag = tag.getCompound(ENTITY_TAG.NBT);
+            entityTag = tag.getCompound(ENTITY_TAG.NBT).copy();
         }
     }
 
