@@ -1,6 +1,7 @@
 package org.bukkit.block.banner;
 
 import junit.framework.Assert;
+import net.minecraft.core.IRegistry;
 import net.minecraft.world.level.block.entity.EnumBannerPatternType;
 import org.bukkit.support.AbstractTestingBase;
 import org.junit.Test;
@@ -9,7 +10,7 @@ public class PatternTypeTest extends AbstractTestingBase {
 
     @Test
     public void testToBukkit() {
-        for (EnumBannerPatternType nms : EnumBannerPatternType.values()) {
+        for (EnumBannerPatternType nms : IRegistry.BANNER_PATTERN) {
             PatternType bukkit = PatternType.getByIdentifier(nms.getHashname());
 
             Assert.assertNotNull("No Bukkit banner for " + nms + " " + nms.getHashname(), bukkit);
@@ -20,7 +21,7 @@ public class PatternTypeTest extends AbstractTestingBase {
     public void testToNMS() {
         for (PatternType bukkit : PatternType.values()) {
             EnumBannerPatternType found = null;
-            for (EnumBannerPatternType nms : EnumBannerPatternType.values()) {
+            for (EnumBannerPatternType nms : IRegistry.BANNER_PATTERN) {
                 if (bukkit.getIdentifier().equals(nms.getHashname())) {
                     found = nms;
                     break;
