@@ -1,9 +1,10 @@
 package org.bukkit.craftbukkit.util;
 
 import static org.junit.Assert.*;
-import net.minecraft.network.chat.ChatComponentText;
+import net.minecraft.network.chat.ComponentContents;
 import net.minecraft.network.chat.IChatBaseComponent;
 import net.minecraft.network.chat.IChatMutableComponent;
+import net.minecraft.network.chat.contents.LiteralContents;
 import org.junit.Test;
 
 public class CraftChatMessageTest {
@@ -96,7 +97,7 @@ public class CraftChatMessageTest {
 
     private boolean containsNonPlainComponent(IChatBaseComponent component) {
         for (IChatBaseComponent c : component) {
-            if (!(c instanceof ChatComponentText)) {
+            if (c.getContents() != ComponentContents.EMPTY && !(c.getContents() instanceof LiteralContents)) {
                 return true;
             }
         }
